Fix warning C4838 - 'conversion requires a narrowing conversion'

+++ src/tests/gl_tests/D3D11EmulatedIndexedBufferTest.cpp
@@ -125,7 +125,9 @@ class D3D11EmulatedIndexedBufferTest : public ANGLETest
 // into a valid emulated indexed buffer.
 TEST_P(D3D11EmulatedIndexedBufferTest, TestNativeToExpandedUsingGLubyteIndices)
 {
-    rx::SourceIndexData srcData = {nullptr, mubyteIndices.data(), mubyteIndices.size(), GL_UNSIGNED_BYTE, false};
+    rx::SourceIndexData srcData = {nullptr, mubyteIndices.data(),
+                                   static_cast<unsigned int>(mubyteIndices.size()),
+                                   GL_UNSIGNED_BYTE, false};
     emulateAndCompare(&srcData);
 }
 
@@ -133,7 +135,9 @@ TEST_P(D3D11EmulatedIndexedBufferTest, TestNativeToExpandedUsingGLubyteIndices)
 // into a valid emulated indexed buffer.
 TEST_P(D3D11EmulatedIndexedBufferTest, TestNativeToExpandedUsingGLushortIndices)
 {
-    rx::SourceIndexData srcData = {nullptr, mushortIndices.data(), mushortIndices.size(), GL_UNSIGNED_SHORT, false };
+    rx::SourceIndexData srcData = {nullptr, mushortIndices.data(),
+                                   static_cast<unsigned int>(mushortIndices.size()),
+                                   GL_UNSIGNED_SHORT, false};
     emulateAndCompare(&srcData);
 }
 
@@ -141,7 +145,9 @@ TEST_P(D3D11EmulatedIndexedBufferTest, TestNativeToExpandedUsingGLushortIndices)
 // into a valid emulated indexed buffer.
 TEST_P(D3D11EmulatedIndexedBufferTest, TestNativeToExpandedUsingGLuintIndices)
 {
-    rx::SourceIndexData srcData = {nullptr, muintIndices.data(), muintIndices.size(), GL_UNSIGNED_INT, false };
+    rx::SourceIndexData srcData = {nullptr, muintIndices.data(),
+                                   static_cast<unsigned int>(muintIndices.size()), GL_UNSIGNED_INT,
+                                   false};
     emulateAndCompare(&srcData);
 }
 
@@ -153,7 +159,9 @@ TEST_P(D3D11EmulatedIndexedBufferTest, TestSourceBufferRemainsUntouchedAfterExpa
     cleanSourceBuffer->copySubData(mSourceBuffer, 0, 0, mSourceBuffer->getSize());
 
     // Do a basic exanded and compare test.
-    rx::SourceIndexData srcData = {nullptr, muintIndices.data(), muintIndices.size(), GL_UNSIGNED_INT, false };
+    rx::SourceIndexData srcData = {nullptr, muintIndices.data(),
+                                   static_cast<unsigned int>(muintIndices.size()), GL_UNSIGNED_INT,
+                                   false};
     emulateAndCompare(&srcData);
 
     const uint8_t *sourceBufferMem = nullptr;
